requirements.txt
Flask==3.0.3
Flask_Cors==4.0.1
openai==0.28.0
python-dotenv==1.0.1
tiktoken==0.5.1